+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features / Enhancements
 
 - Add Checkbox List element (#382)
+- Update syntax support in Code Editor (#383)
 
 ## 3.7.0 (2024-03-10)